Retrieve bodies for rerun transactions OHM-765
